Switch to SoundCache
Use SoundCache from audioplayers it's simpler to setup and use and by
making 'player' nullable I can avoid plugin problems during widget
tests by not having it setup so it doesn't get called. Maybe that would
be an easy way to totally disable sound.

Although the timing is off on the emulator running on a real android
device seems to have reasonable timing.