Use initialId in RoutineSelectorWidget
If there isn't a current routine set in routine provider initialize to
the value from initial id provider which, if it exists, and matches an
existing routine, will trigger a change and have the current routine
setup without build problems.

This makes the most sense anyway since the routines provider is/will be
responsible for handling the persistence of the routines (someday).